requirements.txt
asgiref==3.11.1
async-timeout==4.0.2
boto3==1.23.0
botocore==1.26.1
cachetools==5.0.0
certifi==2024.7.4
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==42.0.4
Deprecated==1.2.13
dj-database-url==0.5.0
Django==4.2.28
django-braces==1.15.0
django-csp==3.7
django-redis==5.2.0
django-storages==1.12.3
gitdb==4.0.9
GitPython==3.1.41
google-api-core==2.7.3
google-api-python-client==2.46.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.0
gunicorn==22.0.0
httplib2==0.20.4
idna==3.7
importlib-metadata==4.11.3
jmespath==1.0.0
mailjet-rest==1.3.4
Markdown==3.3.6
oauth2client==4.1.3
oauthlib==3.2.0
packaging==21.3
ply==3.11
protobuf==3.20.1
psycopg2-binary==2.9.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyotp==2.6.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-decouple==3.6
redis==4.4.4
requests==2.32.2
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.5.2
sentry-sdk==2.13.0
six==1.16.0
smmap==5.0.0
sqlparse==0.5.0
stone==3.3.1
uritemplate==4.1.1
urllib3==1.26.19
whitenoise==6.0.0
wrapt==1.14.1
zipp==3.19.1
# Mon Feb  2 10:15:15 IST 2026
django-ckeditor==6.7.3
django-js-asset==2.2.0
django-jazzmin==3.0.1

# Testing
django-test-enforcer==0.1.0
pytest==8.0.0
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-playwright==0.4.4
playwright==1.40.0
